requirements.txt
pandas>=2.0.3
google-cloud-bigquery>=3.0.0
requests>=2.28.0
python-dotenv>=1.0.0